+++ app/src/main/java/uk/openvk/android/legacy/user_interface/activities/PhotoViewerActivity.java
@@ -134,7 +134,7 @@ public class PhotoViewerActivity extends Activity {
                 try {
                     if (extras.containsKey("original_link") && extras.getString("original_link").length() > 0) {
                         downloadManager = new DownloadManager(this, true);
-                        downloadManager.downloadOnePhotoToCache(extras.getString("original_link"), String.format("original_photo_a%d_%d", extras.getInt("author_id"), extras.getLong("photo_id")), "original_photos");
+                        downloadManager.downloadOnePhotoToCache(extras.getString("original_link"), String.format("original_photo_a%d_%d", extras.getLong("author_id"), extras.getLong("photo_id")), "original_photos");
                     } else {
                         finish();
                     }
@@ -196,7 +196,7 @@ public class PhotoViewerActivity extends Activity {
             bfOptions.inPreferredConfig = Bitmap.Config.ARGB_8888;
             try {
                 Bundle extras = getIntent().getExtras();
-                bitmap = BitmapFactory.decodeFile(String.format("%s/original_photos/original_photo_a%d_%d", getCacheDir().getAbsolutePath(), extras.getInt("author_id"), extras.getLong("photo_id")), bfOptions);
+                bitmap = BitmapFactory.decodeFile(String.format("%s/original_photos/original_photo_a%d_%d", getCacheDir().getAbsolutePath(), extras.getLong("author_id"), extras.getLong("photo_id")), bfOptions);
                 ((ZoomableImageView) findViewById(R.id.picture_view)).setImageBitmap(bitmap);
                 ((ZoomableImageView) findViewById(R.id.picture_view)).enablePinchToZoom();
                 ((ZoomableImageView) findViewById(R.id.picture_view)).setVisibility(View.VISIBLE);
@@ -286,7 +286,7 @@ public class PhotoViewerActivity extends Activity {
     private void savePhoto() {
         Global global = new Global();
         final Bundle data = getIntent().getExtras();
-        String cache_path = String.format("%s/original_photos/original_photo_a%d_%d", getCacheDir().getAbsolutePath(), getIntent().getExtras().getInt("author_id"), getIntent().getExtras().getLong("photo_id"));
+        String cache_path = String.format("%s/original_photos/original_photo_a%d_%d", getCacheDir().getAbsolutePath(), getIntent().getExtras().getLong("author_id"), getIntent().getExtras().getLong("photo_id"));
         File file = new File(cache_path);
         String[] path_array = cache_path.split("/");
         String dest = String.format("%s/OpenVK/Photos/%s", Environment.getExternalStorageDirectory().getAbsolutePath(), path_array[path_array.length - 1]);
